adds basic test configuration
ava was chosen as a test runner (well-maintained, excellent typescript support)
adding this config required slightlying changing the build output config too.
the distributable artifacts are now one level deeper, so that they can remain
in a separate directory from the test artifacts. for clarity, the output directory
was renamed from `dist` to build`, and within build, there is `src` and `test`.
only `build/src` is meant for packaging (and that is reflected in the "main"
property of package.json).