Add default args for 'list' operations
List users and list databases don't require and arguments and the
examples in the documentation don't provide any arguments, for example:

`notion.users.list()`

However, when using this function in TypeScript, this example produces an
error because `list()` requires an argument. This change sets a default
argument of an empty object (`{}`) so the examples in the documentation
work when using TypeScript and an argument is no longer needed.