Remove legacy (<=2021) v1 API shape for page properties (#580)
This PR removes support for the legacy (pre-Q3-2021) shape for page
`properties` parameters from `CreatePageBodyParameters` and
`UpdatePageBodyParameters` in `src/api-endpoints.ts`.

This collapses the union type of {old properties shape, new properties
shape} down to the new shape only.

In the old shape, `properties` was a fairly ambiguous `Record<string,
...>` mapping each property key to the property value (a union type
across all possible property types' values) e.g.

```ts
{
  // ...
  "numberId": 24,
}
```

In the new shape (the only one we want to support going forward), the
value must be wrapped in an object that uses the polymorphic `type`
pattern with the actual value defined in a sub-object to disambiguate:

```ts
{
  // ...
  "numberId": {
    "type": "number",
    "number": 24
  }
}
```

This matches up with our responses in the API read path so most, if not
all, integrations are likely already using syntax compatible with the
new shape, but we'll probably have to cut a new major version for the JS
SDK following this PR.