Fix increment-version workflow to create a PR (#629)
## Description

- **Fix the Git mechanics of the bump version workflow**, since
currently it hits the `main` protected branch rule
[issue](https://github.com/makenotion/notion-sdk-js/actions/runs/18478122676/job/52647175500)
trying to commit directly. Instead, we should **create a PR** and assign
it to whoever triggered the version bump workflow.
- **Add a step to the automatic NPM publish workflow to automatically
push tags for new versions**, so we get the new `v<#>` tags as soon as
the PR created for version bumping is merged to `main`.

## How was this change tested?

- [ ] Automated test (unit, integration, etc.)
- [x] Manual test (provide reproducible testing steps below)

Will manually validate these changes after merging this PR, and iterate
forward on any issues, for simplicity (to avoid confusion from running
these workflows on non-`main` branches which wouldn't be the usual
workflow.)

## Screenshots

N/A