Add support for the place property in the public API (#637)
## Description

Title. Adds support for the new "place" property (see
[here](https://www.notion.com/help/maps) for more) in the public API.
Supports all fields:

```
type PlacePropertyValueResponse = {
  lat: number
  lon: number
  name?: string | null
  address?: string | null
  aws_place_id?: string | null
  google_place_id?: string | null
}
```

`lat` and `lon` are the source-of-truth. The rest of the fields are
additional metadata, so the caller has the responsibility of resolving
the appropriate `lat` and `lon` values (from, say, an address).
`aws_place_id` and `google_place_id` are mostly unused; they supportive
metadata that helps us go from a user's search term to a specific place
ID. If written to via the API, they won't impact anything. They are
included in both the read and write paths to support duplicating values.

## How was this change tested?

- [x] Automated test (unit, integration, etc.)
- [ ] Manual test (provide reproducible testing steps below)

## Screenshots

---------

Co-authored-by: Kenneth Sinder <kenneth@makenotion.com>